Try not to rebuild the sandbox for the benchmark

+++ bench/build.sh
@@ -2,9 +2,12 @@
 
 set -e
 
-cabal sandbox init
-cabal sandbox add-source ../
-cabal install --only-dep
-cabal configure
+if [ ! -f cabal.sandbox.config ]; then
+	cabal sandbox init
+	cabal sandbox add-source ../
+	cabal install --only-dep
+	cabal configure
+fi
+
 cabal build
 ./dist/build/cereal-bench/cereal-bench
